backend/requirements.txt
pandas>=1.5.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-cors>=3.0.0 